Initial stab at mailer opt in/out route
